fixed batch
