+++ internal/fs/copy.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"net/http"
 	stdpath "path"
-	"sync"
 	"time"
 
 	"github.com/OpenListTeam/OpenList/v4/internal/conf"
@@ -63,140 +62,21 @@ func (t *CopyTask) Run() error {
 	taskID := fmt.Sprintf("%p", t)
 	
 	// Register task to batch tracker
-	batchTracker.registerTask(taskID, t.dstStorage, t.DstDirPath)
+	copyBatchTracker.RegisterTask(taskID, t.dstStorage, t.DstDirPath)
 	
 	// Execute copy operation
 	err = copyBetween2Storages(t, t.srcStorage, t.dstStorage, t.SrcObjPath, t.DstDirPath)
 	
-	// Mark task completed and check if cache refresh is needed
-	if err == nil {
-		shouldRefresh, dstStorage, dstDirPath := batchTracker.markTaskCompleted(taskID)
-		if shouldRefresh {
-			op.ClearCache(dstStorage, dstDirPath)
-		}
-	} else {
-		// Clean up tracker records even if task failed
-		batchTracker.markTaskCompleted(taskID)
-	}
+	// Mark task completed and automatically refresh cache if needed
+	copyBatchTracker.MarkTaskCompletedWithRefresh(taskID)
 	
 	return err
 }
 
 var CopyTaskManager *tache.Manager[*CopyTask]
 
-// Batch copy task tracker - aggregates all copy tasks by target directory
-type batchCopyTracker struct {
-	mu           sync.Mutex
-	dirTasks     map[string]*dirTaskInfo // dstStoragePath+dstDirPath -> dirTaskInfo
-	pendingTasks map[string]string       // taskID -> dstStoragePath+dstDirPath
-	lastCleanup  time.Time               // last cleanup time
-}
-
-type dirTaskInfo struct {
-	dstStorage     driver.Driver
-	dstDirPath     string
-	pendingTasks   map[string]bool // taskID -> true
-	lastActivity   time.Time       // last activity time (used for detecting abnormal situations)
-}
-
-var batchTracker = &batchCopyTracker{
-	dirTasks:     make(map[string]*dirTaskInfo),
-	pendingTasks: make(map[string]string),
-	lastCleanup:  time.Now(),
-}
-
-// Generate unique key for target directory
-func (bt *batchCopyTracker) getDirKey(dstStorage driver.Driver, dstDirPath string) string {
-	return dstStorage.GetStorage().MountPath + ":" + dstDirPath
-}
-
-// Register copy task to target directory
-func (bt *batchCopyTracker) registerTask(taskID string, dstStorage driver.Driver, dstDirPath string) {
-	bt.mu.Lock()
-	defer bt.mu.Unlock()
-	
-	// Periodically clean up expired entries
-	bt.cleanupIfNeeded()
-	
-	dirKey := bt.getDirKey(dstStorage, dstDirPath)
-	
-	// Record task to directory mapping
-	bt.pendingTasks[taskID] = dirKey
-	
-	// Initialize or update directory task information
-	if info, exists := bt.dirTasks[dirKey]; exists {
-		info.pendingTasks[taskID] = true
-		info.lastActivity = time.Now()
-	} else {
-		bt.dirTasks[dirKey] = &dirTaskInfo{
-			dstStorage:   dstStorage,
-			dstDirPath:   dstDirPath,
-			pendingTasks: map[string]bool{taskID: true},
-			lastActivity: time.Now(),
-		}
-	}
-}
-
-// Mark task completed, return whether cache refresh is needed
-func (bt *batchCopyTracker) markTaskCompleted(taskID string) (bool, driver.Driver, string) {
-	bt.mu.Lock()
-	defer bt.mu.Unlock()
-	
-	dirKey, exists := bt.pendingTasks[taskID]
-	if !exists {
-		return false, nil, ""
-	}
-	
-	// Remove from pending tasks
-	delete(bt.pendingTasks, taskID)
-	
-	info, exists := bt.dirTasks[dirKey]
-	if !exists {
-		return false, nil, ""
-	}
-	
-	// Remove from directory tasks
-	delete(info.pendingTasks, taskID)
-	
-	// If no pending tasks left in this directory, trigger cache refresh
-	if len(info.pendingTasks) == 0 {
-		dstStorage := info.dstStorage
-		dstDirPath := info.dstDirPath
-		delete(bt.dirTasks, dirKey)  // Delete directly, no need to update lastActivity
-		return true, dstStorage, dstDirPath
-	}
-	
-	// Only update lastActivity when there are other tasks (indicating the directory still has active tasks)
-	info.lastActivity = time.Now()
-	return false, nil, ""
-}
-
-// Check if cleanup is needed, execute cleanup if necessary
-func (bt *batchCopyTracker) cleanupIfNeeded() {
-	now := time.Now()
-	// Clean up every 10 minutes
-	if now.Sub(bt.lastCleanup) > 10*time.Minute {
-		bt.cleanupStaleEntries()
-		bt.lastCleanup = now
-	}
-}
-
-// Clean up timed-out tasks (prevent memory leaks)
-// Mainly used to clean up residual entries caused by abnormal situations (such as task crashes, process restarts, etc.)
-func (bt *batchCopyTracker) cleanupStaleEntries() {
-	now := time.Now()
-	for dirKey, info := range bt.dirTasks {
-		// If no activity for more than 1 hour, it may indicate an abnormal situation, clean up this entry
-		// Under normal circumstances, markTaskCompleted will be called when the task is completed and the entire entry will be deleted
-		if now.Sub(info.lastActivity) > time.Hour {
-			// Clean up related pending tasks
-			for taskID := range info.pendingTasks {
-				delete(bt.pendingTasks, taskID)
-			}
-			delete(bt.dirTasks, dirKey)
-		}
-	}
-}
+// Batch tracker for copy operations
+var copyBatchTracker = NewBatchTracker("copy")
 
 // Copy if in the same storage, call move method
 // if not, add copy task
